Strip form code too
